Link argument parser only in Run
